separate in different files dateutils , apiutils change the id button for a better one change the parameter for calendar , send the context instead mainactivity
