+++ app/src/main/java/cl/mario/covid/ui/main/MainActivity.kt
@@ -42,7 +42,7 @@ class MainActivity : AppCompatActivity() {
 
     private fun initListener(){
         binding.apply {
-            button.setOnClickListener {
+            chooseDateDialogBtn.setOnClickListener {
                 openCalendar()
             }
         }
