+++ app/src/main/res/layout/activity_main.xml
@@ -74,7 +74,7 @@
 
 
             <Button
-                android:id="@+id/button"
+                android:id="@+id/chooseDateDialogBtn"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:text="@string/choose_date"
